rename package from solanadvp to solana.dvp
